fix(db): fix migrations
