--- backend/app/migration/versions/665a1d497755_recreate_lost_migration.py
@@ -1,32 +1,0 @@
-"""recreate lost migration
-
-Revision ID: 665a1d497755
-Revises: 9873e5ab5d76
-Create Date: 2025-05-10 14:23:26.374154
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '665a1d497755'
-down_revision: Union[str, None] = '9873e5ab5d76'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('users_yandex_id_key', 'users', type_='unique')
-    op.drop_column('users', 'yandex_id')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('yandex_id', sa.VARCHAR(length=255), autoincrement=False, nullable=True))
-    op.create_unique_constraint('users_yandex_id_key', 'users', ['yandex_id'])
-    # ### end Alembic commands ###
